+++ src/components/SearchBar.tsx
@@ -11,7 +11,7 @@ import ListItemText from "@mui/material/ListItemText";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
 import Button from "@mui/material/Button";
-import { CardMedia, createTheme, ThemeProvider } from "@mui/material";
+import { createTheme, ThemeProvider } from "@mui/material";
 import { Link } from "react-router-dom";
 import PDLLogo from "./PDLLogo";