+++ src/components/title1/Title1TreeMap.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as d3 from "d3";
 import styled from "styled-components";
-import { Checkbox, FormControlLabel, FormGroup, Grid, IconButton, SvgIcon, Typography } from "@mui/material";
+import { Checkbox, FormControlLabel, FormGroup, Grid, IconButton, Typography } from "@mui/material";
 import SortIcon from "@mui/icons-material/Sort";
 import TreeMapSquares from "./TreeMapSquares";
 import { DownloadIcon } from "../shared/DownloadIcon";
@@ -187,7 +187,7 @@ export default function Title1TreeMap({ program, TreeMapData, year, stateCodes, 
                 .attr("fill", recipientsColor);
         }
     }
-    /* eslint-disable */
+
     return (
         <Styles>
             <Grid
@@ -409,5 +409,4 @@ export default function Title1TreeMap({ program, TreeMapData, year, stateCodes, 
             </Grid>
         </Styles>
     );
-    /* eslint-enable */
 }